+++ esphome/components/ota/ota_backend_arduino_esp32.h
@@ -1,5 +1,6 @@
 #include "esphome/core/defines.h"
-#include "esphome/components/esp8266/preferences.h"
+#if defined(USE_ARDUINO) && defined(USE_ESP32)
+
 #include "ota_component.h"
 #include "ota_backend.h"
 
@@ -7,32 +8,27 @@
 #include <MD5Builder.h>
 #endif
 
-#include <Updater.h>
+#include <Update.h>
 
 namespace esphome {
 namespace ota {
 
-class ArduinoESP8266OTABackend : public OTABackend {
+class ArduinoESP32OTABackend : public OTABackend {
  public:
   OTAResponseTypes begin(size_t image_size) override {
     bool ret = Update.begin(image_size, U_FLASH);
     if (ret) {
-      esp8266::preferences_prevent_write(true);
+      return OTA_RESPONSE_OK;
     }
 
     uint8_t error = Update.getError();
-    if (error == UPDATE_ERROR_BOOTSTRAP)
-      return OTA_RESPONSE_ERROR_INVALID_BOOTSTRAPPING;
-    if (error == UPDATE_ERROR_NEW_FLASH_CONFIG)
-      return OTA_RESPONSE_ERROR_WRONG_NEW_FLASH_CONFIG;
-    if (error == UPDATE_ERROR_FLASH_CONFIG)
-      return OTA_RESPONSE_ERROR_WRONG_CURRENT_FLASH_CONFIG;
-    if (error == UPDATE_ERROR_SPACE)
-      return OTA_RESPONSE_ERROR_ESP8266_NOT_ENOUGH_SPACE;
+    if (error == UPDATE_ERROR_SIZE)
+      return OTA_RESPONSE_ERROR_ESP32_NOT_ENOUGH_SPACE;
     return OTA_RESPONSE_ERROR_UNKNOWN;
   }
 
   void set_update_md5(const char *md5) override { Update.setMD5(md5); }
+
   OTAResponseTypes write(uint8_t *data, size_t len) override {
     size_t written = Update.write(data, len);
     if (written != len) {
@@ -48,10 +44,10 @@ class ArduinoESP8266OTABackend : public OTABackend {
   }
 
   void abort() override {
-    Update.end();
-    esp8266::preferences_prevent_write(false);
+    Update.abort();
   }
 };
 
 }  // namespace ota
 }  // namespace esphome
+#endif
